Refactored with logger class
