+++ requirements.txt
@@ -1,13 +1,94 @@
+aioftp==0.20.1
+aiohttp==3.9.3
+aiohttp-apispec==3.0.0b2
+aiohttp-jinja2==1.5.1
+aiohttp-security==0.4.0
+aiohttp-session==2.12.0
+aiosignal==1.3.1
+alabaster==0.7.16
+apispec==6.4.0
+async-timeout==4.0.3
+asyncio==3.4.3
+asyncssh==2.14.1
+attrs==23.1.0
+Babel==2.14.0
+beautifulsoup4==4.12.2
+bs4==0.0.1
+certifi==2023.11.17
+cffi==1.16.0
+cfgv==3.4.0
+charset-normalizer==3.3.2
+cmd2==2.4.3
+colorama==0.4.6
+colorlog==6.8.0
+cryptography==42.0.2
+cssselect2==0.7.0
+dat-terminal==0.0.1b0
+dirhash==0.2.1
+distlib==0.3.8
+dnspython==2.4.2
+docutils==0.18.1
+filelock==3.13.1
+frozenlist==1.4.0
+identify==2.5.33
+idna==3.4
+imagesize==1.4.1
+IPy==1.1
+Jinja2==3.1.3
+ldap3==2.9.1
+lxml==4.9.4
+Markdown==3.4.4
 markdown-it-py==3.0.0
+MarkupSafe==2.1.5
+marshmallow==3.20.1
+marshmallow-enum==1.5.1
+mdit-py-plugins==0.4.0
 mdurl==0.1.2
-Pygments==2.15.1
+multidict==6.0.4
+myst-parser==2.0.0
+nodeenv==1.8.0
+packaging==23.2
+pathspec==0.12.1
+pillow==10.2.0
+platformdirs==4.1.0
+pre-commit==3.6.0
+pyasn1==0.5.1
+pycparser==2.21
+Pygments==2.17.2
+pyperclip==1.8.2
+PyQt6-Qt6==6.6.3
+PyQt6-sip==13.6.0
+pyreadline3==3.4.1
+PySide6==6.7.0
+PySide6_Addons==6.7.0
+PySide6_Essentials==6.7.0
+PyYAML==6.0.1
+reportlab==4.0.4
+requests==2.31.0
+rich==13.7.0
+scantree==0.0.2
+shiboken6==6.7.0
+snowballstemmer==2.2.0
+soupsieve==2.5
+Sphinx==7.1.2
+sphinx-rtd-theme==1.3.0
+sphinxcontrib-applehelp==1.0.8
+sphinxcontrib-devhelp==1.0.6
+sphinxcontrib-htmlhelp==2.0.5
+sphinxcontrib-jquery==4.1
+sphinxcontrib-jsmath==1.0.1
+sphinxcontrib-qthelp==1.0.7
+sphinxcontrib-serializinghtml==1.1.10
+svglib==1.5.1
+tabulate==0.9.0
 telnetlib3==2.0.2
-aiohttp
-asyncio
-requests>=2.31.0
-rich>=13.6.0
-beautifulsoup4>=4.12.2
-colorama
-tabulate
-cmd2
-IPy
+termcolor==2.4.0
+tinycss2==1.2.1
+typing_extensions==4.10.0
+urllib3==2.1.0
+virtualenv==20.25.0
+wcwidth==0.2.12
+webargs==8.4.0
+webencodings==0.5.1
+websockets==11.0.3
+yarl==1.9.2
