+++ src/logger/__init__.py
@@ -1,2 +1,2 @@
-from .color_formatter import *
 from .logger import *
+import logging
