+++ src/modules/packagemanager.py
@@ -1,6 +1,5 @@
 import subprocess
 from logger import *
-from modules.notify import Notify
 
 
 class PackageManager:
@@ -9,7 +8,6 @@ class PackageManager:
     def __init__(self, program_name):
         self.program_name = program_name
         self.logger = Logger()
-        self.notify = Notify()
 
     def installed(self) -> bool:
         try:
